requirements.txt
env==0.1.0
fastapi==0.111.1
langchain_community==0.2.7
langchain_core==0.2.18
langchain_nomic==0.1.2
langchain_text_splitters==0.2.2
langchain_groq==0.1.6
transformers==4.42.4
pydantic==2.8.2
Requests==2.32.3
streamlit==1.36.0
python-dotenv==1.0.1
pytest==8.3.2